refactor(client): consolidate rackManagement into fleetManagement (#395)
Move pages/, components/, utils/ from `features/rackManagement/` to
`features/fleetManagement/`. Rewrite all `@/protoFleet/features/rackManagement/*`
imports (15 files including routePrefetch.ts factories). Remove now-empty
`features/rackManagement/` directory.
