+++ client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/AssignMinersModal.stories.tsx
