+++ client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/ManageMinersModal.stories.tsx
