+++ client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/ManageMinersModal.test.tsx
