+++ client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/SearchMinersModal.stories.tsx
