+++ client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/index.ts
