+++ client/src/protoFleet/features/fleetManagement/components/RackCard/MiniRackGrid.stories.tsx
