+++ client/src/protoFleet/features/fleetManagement/components/RackCard/RackCard.stories.tsx
