+++ client/src/protoFleet/features/fleetManagement/components/RackDetailGrid/RackDetailSlot.tsx
