+++ client/src/protoFleet/features/fleetManagement/components/RackDetailGrid/index.ts
