+++ client/src/protoFleet/features/fleetManagement/components/RackSlotGrid/RackSlot.tsx
