+++ client/src/protoFleet/features/fleetManagement/components/RackSlotGrid/RackSlotGrid.stories.tsx
