+++ client/src/protoFleet/features/fleetManagement/components/RackSlotGrid/types.ts
@@ -1,4 +1,4 @@
-import type { NumberingOrigin } from "@/protoFleet/features/rackManagement/utils/slotNumbering";
+import type { NumberingOrigin } from "@/protoFleet/features/fleetManagement/utils/slotNumbering";
 
 export type SlotVisualState = "empty" | "occupied" | "selected" | "selectedOccupied" | "dragOver" | "peerHover";
