+++ client/src/protoFleet/features/fleetManagement/pages/RackOverviewPage.tsx
@@ -15,18 +15,18 @@ import { useDeviceSets } from "@/protoFleet/api/useDeviceSets";
 import { useDeviceSetStateCounts } from "@/protoFleet/api/useDeviceSetStateCounts";
 import { useTelemetryMetrics } from "@/protoFleet/api/useTelemetryMetrics";
 import { POLL_INTERVAL_MS } from "@/protoFleet/constants/polling";
-import DeviceSetActionsMenu from "@/protoFleet/features/groupManagement/components/DeviceSetActionsMenu";
-import { DeviceSetPerformanceSection } from "@/protoFleet/features/groupManagement/components/DeviceSetPerformanceSection";
-import FleetErrors from "@/protoFleet/features/kpis/components/FleetErrors";
 import {
   AssignMinersModal,
   type RackFormData,
-} from "@/protoFleet/features/rackManagement/components/AssignMinersModal";
-import SearchMinersModal from "@/protoFleet/features/rackManagement/components/AssignMinersModal/SearchMinersModal";
-import { orderIndexToOrigin } from "@/protoFleet/features/rackManagement/components/AssignMinersModal/types";
-import type { SlotHealthState } from "@/protoFleet/features/rackManagement/components/RackDetailGrid/types";
-import { RackHealthModule } from "@/protoFleet/features/rackManagement/components/RackHealthModule";
-import { SLOT_STATUS_MAP } from "@/protoFleet/features/rackManagement/utils/rackCardMapper";
+} from "@/protoFleet/features/fleetManagement/components/AssignMinersModal";
+import SearchMinersModal from "@/protoFleet/features/fleetManagement/components/AssignMinersModal/SearchMinersModal";
+import { orderIndexToOrigin } from "@/protoFleet/features/fleetManagement/components/AssignMinersModal/types";
+import type { SlotHealthState } from "@/protoFleet/features/fleetManagement/components/RackDetailGrid/types";
+import { RackHealthModule } from "@/protoFleet/features/fleetManagement/components/RackHealthModule";
+import { SLOT_STATUS_MAP } from "@/protoFleet/features/fleetManagement/utils/rackCardMapper";
+import DeviceSetActionsMenu from "@/protoFleet/features/groupManagement/components/DeviceSetActionsMenu";
+import { DeviceSetPerformanceSection } from "@/protoFleet/features/groupManagement/components/DeviceSetPerformanceSection";
+import FleetErrors from "@/protoFleet/features/kpis/components/FleetErrors";
 import { useDuration, useSetDuration } from "@/protoFleet/store";
 import { ChevronDown } from "@/shared/assets/icons";
 import Button, { variants } from "@/shared/components/Button";
