+++ client/src/protoFleet/features/fleetManagement/utils/slotNumbering.ts
