--- client/src/protoFleet/features/rackManagement/index.ts
@@ -1,2 +1,0 @@
-export { default as RackOverviewPage } from "./pages/RackOverviewPage";
-export { default as RacksPage } from "./pages/RacksPage";
